function: Initial commit
Basic framework for functions. For demonstration purposes, turns out
that grub.cfg only uses parameters in the 'gfxmode' function, and
only checks whether it is "keep" or not. Turns out we need it to
*not* be "keep", so we don't need parameters yet.

This is also an example of the Block() interface that is more
complex than if/then/else, as it uses the parser recursively to
implement braces.

Braces aren't yet avaiable outside functions. This is an example
of how to do it, and could be made into common code.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>