goes-build: Add new build primitive operations
Add mv() to move a file with logging.

Add filterCommand() to run an external command as a filter, with
its input set to an io.Reader and its output set to an io.Writer.

Add stripBinary() to strip a binary using the installed strip
command, and return the data as a slice.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>